+++ test/Storage/Adapter/ZendServerDiskTest.php
@@ -1,26 +1,25 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-cache for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-cache/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Cache\Storage\Adapter;
+namespace LaminasTest\Cache\Storage\Adapter;
 
-use Zend\Cache;
-use Zend\Cache\Exception;
+use Laminas\Cache;
+use Laminas\Cache\Exception;
 
 /**
- * @group      Zend_Cache
+ * @group      Laminas_Cache
  */
 class ZendServerDiskTest extends CommonAdapterTest
 {
     public function setUp()
     {
-        if (!getenv('TESTS_ZEND_CACHE_ZEND_SERVER_ENABLED')) {
-            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_ZEND_SERVER_ENABLED to run this test');
+        if (!getenv('TESTS_LAMINAS_CACHE_ZEND_SERVER_ENABLED')) {
+            $this->markTestSkipped('Enable TESTS_LAMINAS_CACHE_ZEND_SERVER_ENABLED to run this test');
         }
 
         if (!function_exists('zend_disk_cache_store') || PHP_SAPI == 'cli') {